+++ lib/monolens/core.rb
@@ -5,6 +5,11 @@ module Monolens
     end
     module_function :chain
 
+    def dig(options)
+      Dig.new(options)
+    end
+    module_function :dig
+
     def mapping(options)
       Mapping.new(options)
     end
@@ -14,4 +19,5 @@ module Monolens
   end
 end
 require_relative 'core/chain'
+require_relative 'core/dig'
 require_relative 'core/mapping'